+++ Desktop/Outubro/Orientação_objeto/24 de outubro/database.py
@@ -1,0 +1,23 @@
+import mysql.connector
+from mysql.connector import Error
+
+class bd:
+    def __init__(self, host, user, password, database):
+        self.host = host
+        self.user = user 
+        self.password = password
+        self.database = database
+
+    def conectar(self):         
+        self.conexao = mysql.connector.connect(
+            host= self.host,
+            user=self.user,
+            password=self.password,
+            database=self.database
+        )
+        self.cursor = self.conexao.cursor()
+        
+    def desconectar(self):
+        self.conexao.close()
+        
+
